+++ Remote Habits/AppDelegate.swift
@@ -8,6 +8,7 @@ import UserNotifications
 @main
 class AppDelegate: UIResponder, UIApplicationDelegate {
     var userManager = DI.shared.userManager
+    let logger = DI.shared.logger
 
     func application(_ application: UIApplication,
                      didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?) -> Bool {
@@ -32,13 +33,16 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
         let configHandler = overrideConfigHandler ?? { config in
             config.logLevel = .debug
             config.autoTrackScreenViews = true
+            if !Env.customerIOTrackURL.isEmpty {
+                config.trackingApiUrl = Env.customerIOTrackURL
+            }
         }
 
         // Step 1: Initialise CustomerIO SDK
         CustomerIO.initialize(siteId: workspaceId, apiKey: apiKey, region: Region.US, configure: configHandler)
 
         // Initialize in-app module
-        MessagingInApp.initialize(organizationId: Env.customerIOInAppOrganizationId)
+        MessagingInApp.initialize(eventListener: self)
     }
 
     // MARK: UISceneSession Lifecycle
@@ -90,5 +94,31 @@ extension AppDelegate: UNUserNotificationCenterDelegate {
                                 withCompletionHandler completionHandler: @escaping (UNNotificationPresentationOptions)
                                     -> Void) {
         completionHandler([.list, .banner, .badge, .sound])
+    }
+}
+
+extension AppDelegate: InAppEventListener {
+    func messageShown(message: InAppMessage) {
+        CustomerIO.shared.track(name: "inapp shown",
+                                data: ["delivery-id": message.deliveryId ?? "(none)", "message-id": message.messageId])
+    }
+
+    func messageDismissed(message: InAppMessage) {
+        CustomerIO.shared.track(name: "inapp dismissed",
+                                data: ["delivery-id": message.deliveryId ?? "(none)", "message-id": message.messageId])
+    }
+
+    func errorWithMessage(message: InAppMessage) {
+        CustomerIO.shared.track(name: "inapp error",
+                                data: ["delivery-id": message.deliveryId ?? "(none)", "message-id": message.messageId])
+    }
+
+    func messageActionTaken(message: InAppMessage, actionValue: String, actionName: String) {
+        CustomerIO.shared.track(name: "inapp action", data: [
+            "delivery-id": message.deliveryId ?? "(none)",
+            "message-id": message.messageId,
+            "action-value": actionValue,
+            "action-name": actionName
+        ])
     }
 }